Made bot stronger
Bot now prioritises to set a Circle to win. If there is no field where the bot can win, the bot looks if the opponent has a field where the opponent could win. If so, the bot sets a circle on that field to block the opponent. If there is no winning field for the bot and the opponent, the bot sets randomly.